reconstruct code version 2
