+++ Assets/Scripts/ClearCounter.cs
@@ -1,14 +1,14 @@
 using System;
 using UnityEngine;
 
-public class ClearCounter : MonoBehaviour
+public class ClearCounter : KitchenObjectHolder
 {
     [SerializeField] private GameObject selectedCounter;
     [SerializeField] private KitchenObjectSO kitchenObjectSO;
-    [SerializeField] private Transform TopPoint;
+
     [SerializeField] private bool testing;
     [SerializeField] private ClearCounter transferTargetCounter;
-    private KitchenObject kitchenObject;
+
 
     private void Update()
     {
@@ -20,15 +20,14 @@ public class ClearCounter : MonoBehaviour
     }
     public void Interact()
     {
-        if (kitchenObject == null)
+        if (GetComponent<KitchenObject>() == null)
         {
-            kitchenObject = GameObject.Instantiate(kitchenObjectSO.prefab, TopPoint).GetComponent<KitchenObject>();
-            kitchenObject.transform.localPosition = Vector3.zero;
+            KitchenObject kitchenObject = GameObject.Instantiate(kitchenObjectSO.prefab, GetHoldPoint()).GetComponent<KitchenObject>();
+            SetKitchenObject(kitchenObject);
         }
         else
         {
             Debug.LogWarning("已經有了-" + gameObject);
-            Destroy(kitchenObject.gameObject);
         }
         print(this.gameObject + "is interacting......");
 
@@ -41,36 +40,5 @@ public class ClearCounter : MonoBehaviour
     {
         selectedCounter.SetActive(false);
     }
-    public KitchenObject GetKitchenObject()
-    {
-        return kitchenObject;
-    }
-    public void TransferKitchenObject(ClearCounter sourceCounter, ClearCounter targetCounter)
-    {
-        if (sourceCounter.GetKitchenObject() == null)
-        {
-            Debug.LogWarning("沒有東西可以轉移");
-            return;
-        }
-        if (targetCounter.GetKitchenObject() != null)
-        {
-            Debug.LogWarning("目標位置已經有東西了");
-            return;
-        }
-        targetCounter.AddKitchenObject(sourceCounter.GetKitchenObject());
-        sourceCounter.ClearKitchenObject();
-    }
-    public void AddKitchenObject(KitchenObject kitchenObject)
-    {
-        kitchenObject.transform.SetParent(TopPoint);
-        kitchenObject.transform.localPosition = Vector3.zero;
-        this.kitchenObject = kitchenObject;
-    }
-    public void ClearKitchenObject()
-    {
-        if (kitchenObject != null)
-        {
-            this.kitchenObject = null;
-        }
-    }
+
 }
