+++ Assets/Scripts/KitchenObjectHolder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: eba6cfe2f76c4544a9cf0c54ab5864dc
